Merge

+++ docassemble_webapp/docassemble/webapp/update.py
@@ -10,18 +10,39 @@ import re
 import sys
 import shutil
 import time
+import fcntl
 from io import open
 
 from distutils.version import LooseVersion
 if __name__ == "__main__":
     import docassemble.base.config
     docassemble.base.config.load(arguments=sys.argv)
+    if 'initialize' in sys.argv:
+        mode = 'initialize'
+    elif 'check_for_updates' in sys.argv:
+        mode = 'check_for_updates'
+    else:
+        mode = 'initialize'
 
 supervisor_url = os.environ.get('SUPERVISOR_SERVER_URL', None)
 if supervisor_url:
     USING_SUPERVISOR = True
 else:
     USING_SUPERVISOR = False
+
+def fix_fnctl():
+    try:
+        flags = fcntl.fcntl(sys.stdout, fcntl.F_GETFL);
+        fcntl.fcntl(sys.stdout, fcntl.F_SETFL, flags&~os.O_NONBLOCK);
+        sys.stderr.write("fix_fnctl: updated stdout\n")
+    except:
+        pass
+    try:
+        flags = fcntl.fcntl(sys.stderr, fcntl.F_GETFL);
+        fcntl.fcntl(sys.stderr, fcntl.F_SETFL, flags&~os.O_NONBLOCK);
+        sys.stderr.write("fix_fnctl: updated stderr\n")
+    except:
+        pass
 
 def remove_inactive_hosts():
     from docassemble.base.config import hostname
@@ -47,7 +68,7 @@ class DummyPackage(object):
         self.name = name
         self.type = 'pip'
         self.limitation = None
-            
+
 def check_for_updates(doing_startup=False):
     sys.stderr.write("check_for_updates: starting\n")
     from docassemble.base.config import hostname
@@ -163,7 +184,7 @@ def check_for_updates(doing_startup=False):
             changed = True
         if 'pycryptodome' not in here_already:
             sys.stderr.write("check_for_updates: installing pycryptodome\n")
-            install_package(DummyPackage('pycryptodome'))            
+            install_package(DummyPackage('pycryptodome'))
             changed = True
         if 'pdfminer' in here_already:
             sys.stderr.write("check_for_updates: uninstalling pdfminer\n")
@@ -207,7 +228,7 @@ def check_for_updates(doing_startup=False):
     for package in Package.query.filter_by(active=True).all():
         if package.type is not None:
             packages[package.id] = package
-            #print "Found a package " + package.name
+            #print("Found a package " + package.name)
     sys.stderr.write("check_for_updates: 4\n")
     for package in Package.query.filter_by(active=False).all():
         if package.name not in package_by_name:
@@ -257,6 +278,8 @@ def check_for_updates(doing_startup=False):
             continue
         if package.name not in here_already:
             sys.stderr.write("check_for_updates: skipping uninstallation of " + str(package.name) + " because not installed" + "\n")
+            returnval = 1
+            newlog = ''
         else:
             returnval, newlog = uninstall_package(package)
         uninstall_done[package.name] = 1
@@ -264,6 +287,9 @@ def check_for_updates(doing_startup=False):
         if returnval == 0:
             Install.query.filter_by(hostname=hostname, package_id=package.id).delete()
             results[package.name] = 'pip uninstall command returned success code.  See log for details.'
+        elif returnval == 1:
+            Install.query.filter_by(hostname=hostname, package_id=package.id).delete()
+            results[package.name] = 'pip uninstall was not run because the package was not installed.'
         else:
             results[package.name] = 'pip uninstall command returned failure code'
             ok = False
@@ -362,6 +388,8 @@ def add_dependencies(user_id):
     for package in installed_packages:
         if package.key in package_by_name:
             continue
+        if package.key.startswith('mysqlclient') or package.key.startswith('mysql-connector') or package.key.startswith('MySQL-python'):
+            continue
         pip_info = get_pip_info(package.key)
         #sys.stderr.write("Home page of " + str(package.key) + " is " + str(pip_info['Home-page']) + "\n")
         Package.query.filter_by(name=package.key).delete()
@@ -420,17 +448,18 @@ def install_package(package):
     if PY2:
         PACKAGE_DIRECTORY = daconfig.get('packages', '/usr/share/docassemble/local')
     else:
-        PACKAGE_DIRECTORY = daconfig.get('packages', '/usr/share/docassemble/local3.5')
+        PACKAGE_DIRECTORY = daconfig.get('packages', '/usr/share/docassemble/local' + text_type(sys.version_info.major) + '.' + text_type(sys.version_info.minor))
     logfilecontents = ''
     pip_log = tempfile.NamedTemporaryFile()
     temp_dir = tempfile.mkdtemp()
-    use_pip_cache = r.get('da:updatepackage:use_pip_cache')
-    if use_pip_cache is None:
-        disable_pip_cache = False
-    elif int(use_pip_cache):
-        disable_pip_cache = False
-    else:
-        disable_pip_cache = True
+    #use_pip_cache = r.get('da:updatepackage:use_pip_cache')
+    #if use_pip_cache is None:
+    #    disable_pip_cache = False
+    #elif int(use_pip_cache):
+    #    disable_pip_cache = False
+    #else:
+    #    disable_pip_cache = True
+    disable_pip_cache = True
     if package.type == 'zip' and package.upload is not None:
         saved_file = SavedFile(package.upload, extension='zip', fix=True)
         commands = ['pip', 'install']
@@ -472,6 +501,7 @@ def install_package(package):
         returnval = 0
     except subprocess.CalledProcessError as err:
         returnval = err.returncode
+    fix_fnctl()
     sys.stderr.flush()
     sys.stdout.flush()
     time.sleep(4)
@@ -504,6 +534,7 @@ def uninstall_package(package):
         returnval = 0
     except subprocess.CalledProcessError as err:
         returnval = err.returncode
+    fix_fnctl()
     sys.stderr.flush()
     sys.stdout.flush()
     time.sleep(4)
@@ -583,18 +614,28 @@ if __name__ == "__main__":
         from docassemble.webapp.packages.models import Package, Install, PackageAuth
         from docassemble.webapp.daredis import r
         #app.config['SQLALCHEMY_DATABASE_URI'] = docassemble.webapp.database.alchemy_connection_string()
-        update_versions()
-        any_package = Package.query.filter_by(active=True).first()
-        if any_package is None:
-            add_dependencies(1)
+        if mode == 'initialize':
             update_versions()
-        check_for_updates(doing_startup=True)
-        remove_inactive_hosts()
-        from docassemble.base.config import daconfig
-        sys.stderr.write("update: touched wsgi file" + "\n")
-        wsgi_file = daconfig.get('webapp', '/usr/share/docassemble/webapp/docassemble.wsgi')
-        if os.path.isfile(wsgi_file):
-            with open(wsgi_file, 'a'):
-                os.utime(wsgi_file, None)
+            any_package = Package.query.filter_by(active=True).first()
+            if any_package is None:
+                add_dependencies(1)
+                update_versions()
+            check_for_updates(doing_startup=True)
+            remove_inactive_hosts()
+        else:
+            check_for_updates()
+            from docassemble.base.config import daconfig
+            if USING_SUPERVISOR:
+                SUPERVISORCTL = daconfig.get('supervisorctl', 'supervisorctl')
+                container_role = ':' + os.environ.get('CONTAINERROLE', '') + ':'
+                if re.search(r':(web|celery|all):', container_role):
+                    args = [SUPERVISORCTL, '-s', 'http://localhost:9001', 'start', 'reset']
+                    result = subprocess.call(args)
+            else:
+                sys.stderr.write("update: touched wsgi file" + "\n")
+                wsgi_file = daconfig.get('webapp', '/usr/share/docassemble/webapp/docassemble.wsgi')
+                if os.path.isfile(wsgi_file):
+                    with open(wsgi_file, 'a'):
+                        os.utime(wsgi_file, None)
         db.engine.dispose()
     sys.exit(0)
